Separate plugins in their own module, chain cloning task
Set up each plugin in its own separate module, with a file fro the
blueprint and another one for the task if needed.

Move the Github cloning to a celery task, chained to the validation one.
This required changes in the way task_ids are handled.
