+++ goodtablesio/tasks.py
@@ -1,4 +1,5 @@
 import datetime
+import logging
 
 import dataset
 from celery import Celery
@@ -6,8 +7,10 @@ from sqlalchemy.types import DateTime
 from sqlalchemy.dialects.postgresql import JSONB
 from goodtables import Inspector
 
-
 from . import config
+
+
+log = logging.getLogger(__name__)
 
 
 # Module API
@@ -15,9 +18,13 @@ from . import config
 app = Celery('tasks')
 app.config_from_object(config)
 
+# TODO: automate
 
-@app.task(name='goodtableio.tasks.validate')
-def validate(payload):
+app.autodiscover_tasks(['goodtablesio.plugins.github'])
+
+
+@app.task(name='goodtablesio.tasks.validate')
+def validate(payload, task_id=None):
     """Main validation task.
 
     Args:
@@ -68,7 +75,6 @@ def validate(payload):
             ```
 
     """
-
     # Get report
     inspector = Inspector(**payload.pop('config', {}))
     report = inspector.inspect(**payload)
@@ -76,7 +82,7 @@ def validate(payload):
     # Save report
     database = dataset.connect(config.DATABASE_URL)
     row = {
-        'task_id': validate.request.id,
+        'task_id': task_id or validate.request.id,
         'report': report,
         'finished': datetime.datetime.utcnow()
     }
